+++ src/crypto/randomx/jit_compiler_x86.cpp
@@ -301,7 +301,7 @@ namespace randomx {
 		}
 		memcpy(code + epilogueOffset, codeEpilogue, epilogueSize);
 
-		codePosFirst = prologueSize + (hasXOP ? loopLoadXOPSize : loopLoadSize);
+		codePosFirst = static_cast<uint32_t>(prologueSize + (hasXOP ? loopLoadXOPSize : loopLoadSize));
 
 #		ifdef XMRIG_FIX_RYZEN
 		mainLoopBounds.first = code + prologueSize;
@@ -405,7 +405,7 @@ namespace randomx {
 		}
 
 		memcpy(code + superScalarHashOffset, codeSshInit, codeSshInitSize);
-		codePos = superScalarHashOffset + codeSshInitSize;
+		codePos = static_cast<uint32_t>(superScalarHashOffset + codeSshInitSize);
 		for (unsigned j = 0; j < RandomX_CurrentConfig.CacheAccesses; ++j) {
 			SuperscalarProgram& prog = programs[j];
 			uint32_t pos = codePos;
@@ -434,7 +434,7 @@ namespace randomx {
 	}
 
 	void JitCompilerX86::generateProgramPrologue(Program& prog, ProgramConfiguration& pcfg) {
-		codePos = ADDR(randomx_program_prologue_first_load) - ADDR(randomx_program_prologue);
+		codePos = static_cast<uint32_t>(ADDR(randomx_program_prologue_first_load) - ADDR(randomx_program_prologue));
 		*(uint32_t*)(code + codePos + 4) = RandomX_CurrentConfig.ScratchpadL3Mask64_Calculated;
 		*(uint32_t*)(code + codePos + 14) = RandomX_CurrentConfig.ScratchpadL3Mask64_Calculated;
 		if (hasAVX) {
@@ -488,7 +488,7 @@ namespace randomx {
 		codePos += 6;
 		emit(RandomX_CurrentConfig.codePrefetchScratchpadTweaked, RandomX_CurrentConfig.codePrefetchScratchpadTweakedSize, code, codePos);
 		memcpy(code + codePos, codeLoopStore, loopStoreSize);
-		codePos += loopStoreSize;
+		codePos += static_cast<uint32_t>(loopStoreSize);
 
 		if (BranchesWithin32B) {
 			const uint32_t branch_begin = static_cast<uint32_t>(codePos);
@@ -507,7 +507,7 @@ namespace randomx {
 
 		*(uint64_t*)(code + codePos) = 0x850f01eb83ull;
 		codePos += 5;
-		emit32(prologueSize - codePos - 4, code, codePos);
+		emit32(static_cast<uint32_t>(prologueSize - codePos - 4), code, codePos);
 		emitByte(0xe9, code, codePos);
 		emit32(epilogueOffset - codePos - 4, code, codePos);
 	}
@@ -898,12 +898,12 @@ namespace randomx {
 		const uint64_t dst = instr.dst;
 
 		if (src != dst) {
-			genAddressReg<true>(instr, src, p, pos);
+			genAddressReg<true>(instr, static_cast<const uint32_t>(src), p, pos);
 			*(uint64_t*)(p + pos) = 0x0604af0f4cull + (dst << 27);
 			pos += 5;
 		}
 		else {
-			emit32(0x86af0f4c + (dst << 27), p, pos);
+			emit32(static_cast<uint32_t>(0x86af0f4c + (dst << 27)), p, pos);
 			genAddressImm(instr, p, pos);
 		}
 
@@ -950,7 +950,7 @@ namespace randomx {
 		const uint64_t dst = instr.dst;
 
 		if (src != dst) {
-			genAddressReg<false>(instr, src, p, pos);
+			genAddressReg<false>(instr, static_cast<uint32_t>(src), p, pos);
 			*(uint64_t*)(p + pos) = 0x0e24f748c08b49ull + (dst << 16);
 			pos += 7;
 		}
@@ -959,7 +959,7 @@ namespace randomx {
 			pos += 6;
 			genAddressImm(instr, p, pos);
 		}
-		*(uint32_t*)(p + pos) = 0xc28b4c + (dst << 19);
+		*(uint32_t*)(p + pos) = static_cast<uint32_t>(0xc28b4c + (dst << 19));
 		pos += 3;
 
 		registerUsage[dst] = pos;
@@ -974,7 +974,7 @@ namespace randomx {
 		const uint64_t dst = instr.dst;
 
 		if (src != dst) {
-			genAddressReg<false>(instr, src, p, pos);
+			genAddressReg<false>(instr, static_cast<uint32_t>(src), p, pos);
 			*(uint32_t*)(p + pos) = static_cast<uint32_t>(0xC4D08B49 + (dst << 16));
 			*(uint64_t*)(p + pos + 4) = 0x0E04F6FB62ULL + (dst << 27);
 			pos += 9;
@@ -998,7 +998,7 @@ namespace randomx {
 
 		*(uint64_t*)(p + pos) = 0x8b4ce8f749c08b49ull + (dst << 16) + (src << 40);
 		pos += 8;
-		emitByte(0xc2 + 8 * dst, p, pos);
+		emitByte(static_cast<uint8_t>(0xc2 + 8 * dst), p, pos);
 
 		registerUsage[dst] = pos;
 		codePos = pos;
@@ -1012,7 +1012,7 @@ namespace randomx {
 		const uint64_t dst = instr.dst;
 
 		if (src != dst) {
-			genAddressReg<false>(instr, src, p, pos);
+			genAddressReg<false>(instr, static_cast<uint32_t>(src), p, pos);
 			*(uint64_t*)(p + pos) = 0x0e2cf748c08b49ull + (dst << 16);
 			pos += 7;
 		}
@@ -1021,7 +1021,7 @@ namespace randomx {
 			pos += 6;
 			genAddressImm(instr, p, pos);
 		}
-		*(uint32_t*)(p + pos) = 0xc28b4c + (dst << 19);
+		*(uint32_t*)(p + pos) = static_cast<uint32_t>(0xc28b4c + (dst << 19));
 		pos += 3;
 
 		registerUsage[dst] = pos;
@@ -1079,7 +1079,7 @@ namespace randomx {
 		const uint64_t dst = instr.dst;
 
 		if (src != dst) {
-			*(uint32_t*)(p + pos) = 0xc0334d + (((dst << 3) + src) << 16);
+			*(uint32_t*)(p + pos) = static_cast<uint32_t>(0xc0334d + (((dst << 3) + src) << 16));
 			pos += 3;
 		}
 		else {
@@ -1100,11 +1100,11 @@ namespace randomx {
 		const uint64_t dst = instr.dst;
 
 		if (src != dst) {
-			genAddressReg<true>(instr, src, p, pos);
-			emit32(0x0604334c + (dst << 19), p, pos);
+			genAddressReg<true>(instr, static_cast<uint32_t>(src), p, pos);
+			emit32(static_cast<uint32_t>(0x0604334c + (dst << 19)), p, pos);
 		}
 		else {
-			*(uint32_t*)(p + pos) = 0x86334c + (dst << 19);
+			*(uint32_t*)(p + pos) = static_cast<uint32_t>(0x86334c + (dst << 19));
 			pos += 3;
 			genAddressImm(instr, p, pos);
 		}
@@ -1125,7 +1125,7 @@ namespace randomx {
 			pos += 6;
 		}
 		else {
-			*(uint32_t*)(p + pos) = 0xc8c149 + (dst << 16);
+			*(uint32_t*)(p + pos) = static_cast<uint32_t>(0xc8c149 + (dst << 16));
 			pos += 3;
 			emitByte(instr.getImm32() & 63, p, pos);
 		}
@@ -1146,7 +1146,7 @@ namespace randomx {
 			pos += 6;
 		}
 		else {
-			*(uint32_t*)(p + pos) = 0xc0c149 + (dst << 16);
+			*(uint32_t*)(p + pos) = static_cast<uint32_t>(0xc0c149 + (dst << 16));
 			pos += 3;
 			emitByte(instr.getImm32() & 63, p, pos);
 		}
